backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
annotated-types==0.7.0
anthropic==0.96.0
anyio==4.13.0
asgiref==3.11.1
attrs==26.1.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
colorama==0.4.6
coverage==7.13.5
cryptography==46.0.7
distro==1.9.0
Django==4.2
dj-database-url==2.3.0
django-cors-headers==4.9.0
django-filter==25.1
djangorestframework==3.17.1
djangorestframework_simplejwt==5.5.1
docstring_parser==0.18.0
factory_boy==3.3.3
Faker==40.15.0
firebase-admin==6.5.0
frozenlist==1.8.0
google-auth==2.49.2
google-genai==1.55.0
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.12
iniconfig==2.3.0
jiter==0.14.0
multidict==6.7.1
packaging==26.1
pillow==12.2.0
pluggy==1.6.0
propcache==0.4.1
psycopg2-binary==2.9.12
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.13.3
pydantic_core==2.46.3
Pygments==2.20.0
PyJWT==2.12.1
pytest==9.0.3
pytest-django==4.12.0
python-decouple==3.8
requests==2.33.1
sniffio==1.3.1
sqlparse==0.5.5
tenacity==9.1.4
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.1
urllib3==2.6.3
websockets==15.0.1
whitenoise==6.12.0
yarl==1.23.0
